+++ core/src/main/lua/script/builtin/vn/textlog.lua
@@ -20,17 +20,11 @@ end
 
 ---Shows the textlog screen.
 function textLog()
-    local oldContext = getCurrentContext()
-
-    local newContext = createContext(function()
+    callInContext(function()
         local screen = textLogConstructor()
         screen:show()
         screen:destroy()
-        setContextActive(oldContext, true)
     end)
-    setContextActive(newContext, true)
-    setContextActive(oldContext, false)
-    yield()
 end
 
 ---TextLog screen